feat(python): dissolve linalg into System methods + top-level coefficients
PR #64 (part D). linalg was never linear algebra; move its pieces to the
everyday API:

- System gains friendly building methods (add_functions, add_linear,
  add_linear_forms, add_products_of_linears, randomize, add_slices_as_products)
  in _system_ops.py, attached to the bound class. They accept numpy arrays /
  lists and exact coefficient matrices, coerce, and delegate to the low-level
  C++ block methods. The friendly randomize captures the native one first, so
  its optional-matrix override doesn't recurse.
- coefficient(value) / coefficients(iterable) move to the top level
  (bertini.coefficient / bertini.coefficients; as_coefficients renamed,
  dropping the as_ verb), sourced from the neutral _coefficients.py.
- Slice.from_coefficients becomes the friendly, coercing classmethod (exact
  values first, a plain variable list accepted); it captures the native static
  to reach it. Only linalg called the native arg order, so the override is safe.

linalg is now a thin DEPRECATED shim forwarding to these homes, so tests
(linalg_test, slice_witness, mpi) stay green. variable_matrix / multiprec.Vector
remain as deprecated shims; hard-removal + the test sweep ride the cleanup (#23).
VariableGroup(name,count) absorption of variable_vector needs a binding change
and rides the recompile batch (#18/#21).

Full suite: 584 passed (same 4 pre-existing unicode-parser failures).
